move temperature chart completely to cold chain common
