--- client/packages/coldchain/src/Mobile/Monitoring/TemperatureChart.tsx
@@ -1,5 +1,0 @@
-import React from 'react';
-
-export const TemperatureChart = () => {
-  return <div>TemperatureChart</div>;
-};
