--- client/packages/coldchain/src/common/Monitoring/Chart/index.ts
@@ -1,1 +1,0 @@
-export { Chart } from './Chart';
