+++ client/packages/coldchain/src/common/Monitoring/index.ts
@@ -1,1 +1,1 @@
-export * from './Chart';
+export * from './TemperatureChart';
